Fix NullPointerException in gnu.text.CharMap
AbstractWeakHashTable.WEntry.getValue() calls WeakReference.get(),
which will return null if its referent has been GCed. CharMap doesn't
test whether the result of this call is null, so in rare instances a
NullPointerException is thrown when operating over Chars.

This commit adds a check whether the returned value is null before
attempting to deference it to prevent the NPE from occurring.